requirements.txt
keras==2.15.0
langchain==0.0.340
matplotlib==3.6.2
numpy==1.23.5
openai==1.3.5
pandas==2.1.3
Pillow==9.3.0
Pillow==9.0.1
Pillow==10.1.0
scikit_learn==1.3.2
streamlit==1.28.2
tensorflow==2.15.0
tqdm==4.64.1
